Cleanup App.tsx
